requirements.txt
torch
torchvision
pillow
numpy
tqdm
facenet-pytorch
scikit-learn